refactor(workspace): virtualize ConversationListView; extract useWorkspaceData + useConversationPreviews hooks
- ConversationListView: install virtua, use VList for O(1) DOM on large lists,
  extract ConversationItem as memo'd component
- WorkspacePage: 2177 → 1956 lines; matters/contacts/invoices data and
  conversation preview loading moved to dedicated hooks
