+++ package.json
@@ -73,6 +73,7 @@
     "tailwind-merge": "2.6.0",
     "tldts": "^7.0.9",
     "ulid": "^3.0.1",
+    "virtua": "^0.49.1",
     "web-vitals": "^5.2.0",
     "zod": "^4.1.11"
   },
