+++ src/features/chat/pages/WorkspacePage.tsx
@@ -25,7 +25,6 @@ import { Button } from '@/shared/ui/Button';
 import { useConversations } from '@/shared/hooks/useConversations';
 import {
   addConversationTag,
-  fetchLatestConversationMessage,
   getConversation,
   removeConversationTag,
   updateConversationMetadata,
@@ -42,12 +41,11 @@ import {
   shouldShowConversationInPracticeInbox,
 } from '@/shared/utils/conversationDisplay';
 import { resolveConsultationState } from '@/shared/utils/consultationState';
-import { formatLongDate } from '@/shared/utils/dateFormatter';
 import { usePracticeManagement } from '@/shared/hooks/usePracticeManagement';
 import { usePracticeTeam } from '@/shared/hooks/usePracticeTeam';
 import { usePracticeDetails } from '@/shared/hooks/usePracticeDetails';
-import { useMattersData } from '@/shared/hooks/useMattersData';
-import { useClientsData } from '@/shared/hooks/useClientsData';
+import { useWorkspaceData } from './hooks/useWorkspaceData';
+import { useConversationPreviews } from './hooks/useConversationPreviews';
 import { useMessageHandling } from '@/shared/hooks/useMessageHandling';
 import {
   type BasicsFormValues,
@@ -76,11 +74,7 @@ import {
 } from '@/shared/utils/workspaceShell';
 import {
   CLIENT_CONVERSATIONS_ASSIGNED_TO_MAP,
-  CLIENT_INVOICES_FILTER_MAP,
-  CLIENT_MATTERS_FILTER_MAP,
-  MATTERS_FILTER_MAP,
   PRACTICE_CONVERSATIONS_ASSIGNED_TO_MAP,
-  PRACTICE_INVOICES_FILTER_MAP,
   type SecondaryNavItem,
   type WorkspaceSection,
   getClientNavConfig,
@@ -92,7 +86,6 @@ import SecondaryPanel from '@/shared/ui/nav/SecondaryPanel';
 import InspectorPanel from '@/shared/ui/inspector/InspectorPanel';
 import { SettingsContent, type SettingsView } from '@/features/settings/pages/SettingsContent';
 import { mockApps } from '@/features/settings/pages/appsData';
-import { listClientInvoices, listInvoices } from '@/features/invoices/services/invoicesService';
 import type { ChatMessageUI } from '../../../../worker/types';
 import type { Conversation, ConversationMode } from '@/shared/types/conversation';
 import type { LayoutMode } from '@/app/MainApp';
@@ -185,11 +178,6 @@ const hasIntakeContactStarted = (messages: ChatMessageUI[]): boolean => {
     if (meta?.intakeComplete === true) return true;
     return false;
   });
-};
-
-const toBillingTypeLabel = (value?: string | null) => {
-  if (!value) return null;
-  return value.replace(/_/g, ' ').replace(/\b\w/g, (letter) => letter.toUpperCase());
 };
 
 const WorkspacePage: FunctionComponent<WorkspacePageProps> = ({
@@ -631,147 +619,29 @@ const WorkspacePage: FunctionComponent<WorkspacePageProps> = ({
     window.addEventListener('workspace:open-inspector', handleOpenInspector);
     return () => window.removeEventListener('workspace:open-inspector', handleOpenInspector);
   }, [inspectorTarget]);
-  const mattersStatusFilter = useMemo<string[]>(() => {
-    if (workspaceSection !== 'matters') return [];
-    if (!activeSecondaryFilter) return [];
-    if (isPracticeWorkspace) {
-      return MATTERS_FILTER_MAP[activeSecondaryFilter] ?? [];
-    }
-    if (isClientWorkspace) {
-      return CLIENT_MATTERS_FILTER_MAP[activeSecondaryFilter] ?? [];
-    }
-    return [];
-  }, [activeSecondaryFilter, isClientWorkspace, isPracticeWorkspace, workspaceSection]);
-  const contactsStatusFilter = useMemo<UserDetailStatus | null>(() => null, []);
-  const invoicesStatusFilter = useMemo<string[]>(() => {
-    if (workspaceSection !== 'invoices') return [];
-    if (!activeSecondaryFilter) return [];
-    if (isPracticeWorkspace) {
-      return PRACTICE_INVOICES_FILTER_MAP[activeSecondaryFilter] ?? [];
-    }
-    if (isClientWorkspace) {
-      return CLIENT_INVOICES_FILTER_MAP[activeSecondaryFilter] ?? [];
-    }
-    return [];
-  }, [activeSecondaryFilter, isClientWorkspace, isPracticeWorkspace, workspaceSection]);
-  // Always fetch the full unfiltered matters list so the inspector can use it.
-  // The mattersStore handles deduplication — this fires once and caches.
-  const mattersData = useMattersData(
-    practiceId,
-    [], // no status filter — fetch all, filter at display time
-    { enabled: isPracticeWorkspace || isClientWorkspace }
-  );
-  // Filtered view for the matters list page (status filter applied after fetch)
-  const filteredMattersItems = useMemo(() => {
-    if (!mattersStatusFilter || mattersStatusFilter.length === 0) return mattersData.items;
-    const accepted = new Set(mattersStatusFilter.map((s) => s.trim().toLowerCase()));
-    return mattersData.items.filter((m) => accepted.has(String(m.status ?? '').toLowerCase()));
-  }, [mattersData.items, mattersStatusFilter]);
-  const mattersDataForView = useMemo(() => ({
-    ...mattersData,
-    items: filteredMattersItems,
-  }), [mattersData, filteredMattersItems]);
-  const contactsData = useClientsData(
-    practiceId,
+  const {
+    mattersStatusFilter,
     contactsStatusFilter,
+    invoicesStatusFilter,
+    mattersData,
+    mattersDataForView,
+    contactsData,
+    matterClientOptions,
+    matterClientPeople,
+    selectedMatterInspectorData,
+    hasDesktopInvoiceListItems,
+  } = useWorkspaceData({
+    practiceId,
+    isPracticeWorkspace,
+    isClientWorkspace,
+    view,
+    layoutMode,
+    workspaceSection,
+    activeSecondaryFilter,
+    selectedMatterIdFromPath,
     sessionUserId,
-    { enabled: isPracticeWorkspace && (view === 'contacts' || view === 'matters') }
-  );
-  const selectedMatter = useMemo(
-    () => mattersData?.items?.find((matter) => matter.id === selectedMatterIdFromPath) ?? null,
-    [mattersData?.items, selectedMatterIdFromPath]
-  );
-  const matterClientOptions = useMemo<ComboboxOption[]>(
-    () => (contactsData?.items ?? [])
-      .map((client): ComboboxOption | null => {
-        const userId = client.user?.id;
-        if (!userId) return null;
-        return {
-          value: userId,
-          label: (() => {
-            const name = client.user?.name?.trim();
-            return name && name.length ? name : client.user?.email ?? 'Unknown contact';
-          })(),
-          meta: client.user?.email ?? undefined,
-        };
-      })
-      .filter((option): option is ComboboxOption => option !== null),
-    [contactsData?.items]
-  );
-  const matterClientPeople = useMemo(
-    () => (contactsData?.items ?? [])
-      .map((client) => {
-        const userId = client.user?.id;
-        if (!userId) return null;
-        return {
-          userId,
-          name: (() => {
-            const name = client.user?.name?.trim();
-            return name && name.length ? name : client.user?.email ?? 'Unknown contact';
-          })(),
-          email: client.user?.email ?? undefined,
-          image: null,
-          role: 'client',
-        };
-      })
-      .filter((client) => client !== null),
-    [contactsData?.items]
-  );
-  const selectedMatterInspectorData = useMemo(() => {
-    if (!selectedMatter) return null;
-
-    const clientNameById = new Map(
-      (contactsData?.items ?? []).map((client) => [
-        client.user?.id ?? '',
-        client.user?.name ?? client.user?.email ?? '',
-      ])
-    );
-    const clientNameFromId = selectedMatter.client_id ? clientNameById.get(selectedMatter.client_id) : null;
-    const selectedMatterRecord = selectedMatter as Record<string, unknown>;
-    const selectedMatterClientName = clientNameFromId
-      ?? (typeof selectedMatterRecord.client_name === 'string' ? selectedMatterRecord.client_name : null);
-
-    const assigneeNamesFromRows = Array.isArray(selectedMatter.assignees)
-      ? selectedMatter.assignees
-        .map((assignee) => {
-          if (typeof assignee === 'string') {
-            return assignee.trim();
-          }
-          if (!assignee || typeof assignee !== 'object') return '';
-          const row = assignee as Record<string, unknown>;
-          const name = typeof row.name === 'string'
-            ? row.name
-            : (typeof row.email === 'string' ? row.email : '');
-          return name.trim();
-        })
-        .filter((name): name is string => name.length > 0)
-      : [];
-    const selectedMatterAssigneeNames = assigneeNamesFromRows.length > 0
-      ? assigneeNamesFromRows
-      : (selectedMatter.assignee_ids?.map((id) => `User ${id.slice(0, 6)}`) ?? []);
-
-    return {
-      matterClientName: selectedMatterClientName,
-      matterAssigneeNames: selectedMatterAssigneeNames,
-      matterBillingLabel: toBillingTypeLabel(selectedMatter.billing_type),
-      matterCreatedLabel: formatLongDate(selectedMatter.created_at),
-      matterUpdatedLabel: selectedMatter.updated_at
-        ? `Updated ${formatRelativeTime(selectedMatter.updated_at)}`
-        : null,
-      matterClientId: selectedMatter.client_id ?? null,
-      matterUrgency: typeof selectedMatter.urgency === 'string' ? selectedMatter.urgency : null,
-      matterResponsibleAttorneyId: selectedMatter.responsible_attorney_id ?? null,
-      matterOriginatingAttorneyId: selectedMatter.originating_attorney_id ?? null,
-      matterCaseNumber: selectedMatter.case_number ?? null,
-      matterType: selectedMatter.matter_type ?? null,
-      matterCourt: selectedMatter.court ?? null,
-      matterJudge: selectedMatter.judge ?? null,
-      matterOpposingParty: selectedMatter.opposing_party ?? null,
-      matterOpposingCounsel: selectedMatter.opposing_counsel ?? null,
-    };
-  }, [contactsData?.items, selectedMatter]);
+  });
   const isMobileLayout = layoutMode !== 'desktop';
-  const [hasDesktopInvoiceListItems, setHasDesktopInvoiceListItems] = useState<boolean | null>(null);
 
   useEffect(() => {
     onboardingConversationInitRef.current = false;
@@ -780,48 +650,6 @@ const WorkspacePage: FunctionComponent<WorkspacePageProps> = ({
     setOnboardingConversationId(null);
     setOnboardingConversationRetryTick(0);
   }, [practiceId]);
-
-  useEffect(() => {
-    if (layoutMode !== 'desktop' || view !== 'invoices') {
-      setHasDesktopInvoiceListItems(null);
-      return;
-    }
-    if (!practiceId || (!isPracticeWorkspace && !isClientWorkspace)) {
-      setHasDesktopInvoiceListItems(null);
-      return;
-    }
-
-    const controller = new AbortController();
-
-    void (async () => {
-      try {
-        const result = isPracticeWorkspace
-          ? await listInvoices(
-              practiceId,
-              { rules: [], page: 1, pageSize: 1 },
-              { signal: controller.signal, statusFilter: invoicesStatusFilter }
-            )
-          : await listClientInvoices(
-              practiceId,
-              { rules: [], page: 1, pageSize: 1 },
-              { signal: controller.signal, statusFilter: invoicesStatusFilter }
-            );
-
-        if (!controller.signal.aborted) {
-          setHasDesktopInvoiceListItems(result.total > 0);
-        }
-      } catch (error) {
-        if ((error as DOMException)?.name === 'AbortError' || axios.isCancel(error)) {
-          return;
-        }
-        if (!controller.signal.aborted) {
-          setHasDesktopInvoiceListItems(true);
-        }
-      }
-    })();
-
-    return () => controller.abort();
-  }, [invoicesStatusFilter, isClientWorkspace, isPracticeWorkspace, layoutMode, practiceId, view]);
 
   const onboardingConversationFromList = useMemo(() => {
     if (!isPracticeWorkspace) return null;
@@ -903,66 +731,17 @@ const WorkspacePage: FunctionComponent<WorkspacePageProps> = ({
     })();
   }, [createOnboardingConversation, isAnonymous, isConversationsLoading, isPracticeWorkspace, isSessionPending, onboardingConversationFromList, onboardingConversationId, onboardingConversationRetryTick, practiceId, refreshConversations, sessionUserId, view]);
 
-  const [conversationPreviews, setConversationPreviews] = useState<Record<string, {
-    content: string;
-    role: string;
-    createdAt: string;
-  }>>({});
-  const fetchedPreviewIds = useRef<Set<string>>(new Set());
-  const previewFailureCounts = useRef<Record<string, number>>({});
-  const MAX_PREVIEW_ATTEMPTS = 2;
-  const shouldLoadConversationPreviews = view === 'home' || view === 'list' || view === 'conversation';
-
-  useEffect(() => {
-    fetchedPreviewIds.current = new Set();
-    previewFailureCounts.current = {};
-    setConversationPreviews(mockConversationPreviews ?? {});
-  }, [practiceId, mockConversationPreviews, mockConversations]);
-
-  useEffect(() => {
-    if (mockConversationPreviews || mockConversations) return;
-    if (!shouldLoadConversationPreviews || filteredConversations.length === 0 || !practiceId) {
-      return;
-    }
-    if (workspace === 'practice' && (isSessionPending || isAnonymous || !sessionUserId)) {
-      return;
-    }
-    let isMounted = true;
-    const loadPreviews = async () => {
-      const updates: Record<string, { content: string; role: string; createdAt: string }> = {};
-      const toFetch = filteredConversations.slice(0, 10).filter(
-        (conversation) => !fetchedPreviewIds.current.has(conversation.id)
-      );
-      await Promise.all(toFetch.map(async (conversation) => {
-        const message = await fetchLatestConversationMessage(
-          conversation.id,
-          practiceId
-        ).catch(() => null);
-        if (message?.content) {
-          fetchedPreviewIds.current.add(conversation.id);
-          updates[conversation.id] = {
-            content: message.content,
-            role: message.role,
-            createdAt: message.created_at
-          };
-          return;
-        }
-        const currentFailures = previewFailureCounts.current[conversation.id] ?? 0;
-        const nextFailures = currentFailures + 1;
-        previewFailureCounts.current[conversation.id] = nextFailures;
-        if (nextFailures >= MAX_PREVIEW_ATTEMPTS) {
-          fetchedPreviewIds.current.add(conversation.id);
-        }
-      }));
-      if (isMounted && Object.keys(updates).length > 0) {
-        setConversationPreviews((prev) => ({ ...prev, ...updates }));
-      }
-    };
-    void loadPreviews();
-    return () => {
-      isMounted = false;
-    };
-  }, [filteredConversations, isAnonymous, isSessionPending, mockConversationPreviews, mockConversations, practiceId, sessionUserId, shouldLoadConversationPreviews, workspace]);
+  const { conversationPreviews } = useConversationPreviews({
+    practiceId,
+    view,
+    workspace,
+    filteredConversations,
+    isSessionPending,
+    isAnonymous,
+    sessionUserId,
+    mockConversationPreviews,
+    mockConversations,
+  });
 
   const handleSelectConversation = useCallback((conversationId: string) => {
     hasAutoNavigatedRef.current = true;
